nwffn4-multi-tenant-saas-permission-system/repository_before/requirements.txt
Django==4.2.9
djangorestframework==3.14.0
psycopg2-binary==2.9.9
django-redis==5.4.0
redis==5.0.1
